pyproject.toml
[tool.poetry]
name = "scanner-py"
version = "0.1.0"
description = ""
authors = ["Eric Butera <eric@ericbutera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.26.99"
cloudevents = "^1.9.0"
requests = "^2.28.2"
flake8 = "^6.0.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"
asyncio = "^3.4.3"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
